requirements.txt
streamlit
streamlit-authenticator
pandas
numpy
scikit-learn
joblib
plotlys